Wenn in der CacheList ein Mystery Cache mit Final Wegpunkt gewählt wird, gleich den Final aktivieren.
When in CacheList a Mystery with a final waypoint will be selected, select the final waypoint.

CompassView: 
Wenn ein Wegpunkt gewählt ist wird die Richtung und Distanz zu diesem Wegpunkt gezeigt und nicht zur Cache-Koordinate

When a Waypoint is selected the direction and distance to this waypoint will be shown not to the cache coordinates.

git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@102 5890f383-b09f-485f-8f59-22bc967d67fd